pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finamt"
version = "0.17.0"
description = "An agentic Python library for extracting key information from receipts and preparing essential German tax return statements."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Yauheniya Varabyova"}
]
maintainers = [
    {name = "Yauheniya Varabyova"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: German",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
    "Topic :: Multimedia :: Graphics :: Capture :: Digital Camera",
]
keywords = ["finanzamt", "finance", "office", "receipts", "invoices", "tax", "statement", "ocr", "extraction"]
requires-python = ">=3.10,<3.14"
dependencies = [
    "requests>=2.32.0",
    "PyMuPDF>=1.22.0",
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0",
    "pytesseract>=0.3.10",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "typer>=0.12.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.18",
    "sphinx-copybutton>=0.5",
]
dev = [
    "twine",
    "arelle-release",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "finamt[docs]",
]

[project.urls]
"Changelog" = "https://github.com/yauheniya-ai/finamt/blob/main/CHANGELOG.md"
"Documentation" = "https://finamt.readthedocs.io"
"Repository" = "https://github.com/yauheniya-ai/finamt"

[project.scripts]
finamt = "finamt.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["finamt"]
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.setuptools]
include-package-data = true
